Move helper methods from append test to a separate header file
